+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## 0.2.0
+
+Guarantee commands are run in the root project dir or else the custom
+`cwd` project sub-dir defined for the command.
+
 ## 0.1.1
 
 Update project metadata.